+++ cqrs-sample/cqrs-sample-generic-test/src/main/java/com/damon/cqrs/sample/red_packet/api/event/RedPacketGrabSucceedEvent.java
@@ -1,5 +1,7 @@
 package com.damon.cqrs.sample.red_packet.api.event;
 
+import java.math.BigDecimal;
+
 import com.damon.cqrs.domain.Event;
 import lombok.Data;
 
@@ -10,7 +12,7 @@ public class RedPacketGrabSucceedEvent extends Event {
      *
      */
     private static final long serialVersionUID = 8892155336754024236L;
-    private Double money;
+    private BigDecimal money;
     private Long userId;
     private Long redPacketId;
 
@@ -18,7 +20,7 @@ public class RedPacketGrabSucceedEvent extends Event {
         super();
     }
 
-    public RedPacketGrabSucceedEvent(Double money, Long userId) {
+    public RedPacketGrabSucceedEvent(BigDecimal money, Long userId) {
         this.money = money;
         this.userId = userId;
     }
